@@ -27,16 +27,26 @@
       "last 1 safari version"
     ]
   },
-  "dependencies": {
-    "express": "^4.17.3"
-  },
   "devDependencies": {
-    "@testing-library/jest-dom": "^5.16.2",
-    "@testing-library/react": "^12.1.2",
+    "@testing-library/jest-dom": "^5.16.5",
+    "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^13.5.0",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-scripts": "5.0.0",
+    "axios": "^1.2.6",
+    "cors": "^2.8.5",
+    "dotenv": "^16.0.3",
+    "form-data": "^4.0.0",
+    "react": "^18.2.0",
+    "react-datepicker": "^4.10.0",
+    "react-device-detect": "^2.2.2",
+    "react-dom": "^18.2.0",
+    "react-donut-chart": "^1.3.3",
+    "react-minimal-pie-chart": "^8.4.0",
+    "react-router-dom": "^6.8.0",
+    "react-scripts": "5.0.1",
+    "react-zoom-pan-pinch": "^2.2.1",
+    "recharts": "^2.4.1",
+    "semantic-ui-react": "^2.1.4",
+    "tachyons": "^4.12.0",
     "web-vitals": "^2.1.4"
   }
 }
