@@ -1,4 +1,3 @@
-
 const express = require('express')
 const path = require("path");
 const app = express()
@@ -9,7 +8,7 @@ const app = express()
 var options = {
   dotfiles: 'ignore',
   etag: false,
-  extensions: ['htm', 'html','css','js','ico','jpg','jpeg','png','svg'],
+  extensions: ['htm', 'html','css','js','ico','jpg','jpeg','png','svg', 'gif'],
   index: ['index.html'],
   maxAge: '1m',
   redirect: false
