@@ -1,25 +1,17 @@
-import logo from './logo.svg';
+import { Routes, Route } from 'react-router-dom';
 import './App.css';
+import Dashboard from './formular/Dashboard';
+import FormContainer from './formular/FormContainer';
+import Home from './formular/Home';
 
 function App() {
   return (
-    <div className="App">
-      <header className="App-header">
-        <img src={logo} className="App-logo" alt="logo" />
-        <p>
-          Edit <code>src/App.js</code> and save to reload.
-        </p>
-        <a
-          className="App-link"
-          href="https://reactjs.org"
-          target="_blank"
-          rel="noopener noreferrer"
-        >
-          Learn React
-        </a>
-      </header>
-    </div>
+    <Routes>
+        <Route path='/' element={<Home />} />
+        <Route path='/form' element={<FormContainer />} />
+        <Route path='/dashboard' element={<Dashboard />} />
+    </Routes>
   );
-}
+};
 
 export default App;
